refactor(memory): extract date utility functions into date-utils.ts
Move 6 standalone date/time functions out of observational-memory.ts
into a dedicated date-utils.ts module:

- formatRelativeTime
- formatGapBetweenDates
- parseDateFromContent
- isFutureIntentObservation
- expandInlineEstimatedDates
- addRelativeTimeToObservations

No public API changes. First step of OM decomposition.
