+++ .changeset/calm-symbols-feel.md
@@ -1,0 +1,5 @@
+---
+'@mastra/memory': patch
+---
+
+Improved code organization by extracting date utility functions into a dedicated module for better maintainability.
